chore: move injected to a separate export
