+++ packages/vitest/src/index.ts
@@ -15,6 +15,7 @@ export { runOnce, isFirstRun } from './integrations/run-once'
 export * from './integrations/chai'
 export * from './integrations/vi'
 export * from './integrations/utils'
+export { inject } from './integrations/inject'
 export type { SnapshotEnvironment } from '@vitest/snapshot/environment'
 
 export * from './types'
